update ssh-agent initialization
sometimes, the socket isn't removed when the computer turns off. This
leads to a scenario where the socket exists but the agent isn't running.
This commit addresses this by checking if an agent is running if the
socket exists, and if one isn't it deletes the socket and starts the
agent.